requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.2.1
openai==1.58.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
tinytag==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
